Type erased TCB removing the dependency on the TDB definition
The definition of the TCB was previously required, but it wasn't actually used in the scheduler, it's up to the target to cast and manipulate the tcb based on the HAL interface.

The device_fiber.h has been removed, and the definitions of the functions have been added to codal_target_hal.h, removing a further dependency. It's now up to the target to implement all functions in codal_target_hal.h