PktSerial goes low for only 10 microseconds
Some UART hardware get's confused if the low pulse is to close to the beginning of the uart frames. I.e. a low to high can be seen as a start condition.

This commit reduces the low time, and the time between the serial start to 10 microseconds lo, and 100 microseconds high before starting.