BUGFIX: device hang after packet received.
PktSerial did not allocate a fresh buffer when adding to its txQueue (bad design by me). Thus when a packet buffer used in the bridge went out of scope and was destroyed bad things happened.

This commit ensures pktserial allocates its own buffer and memcpy's the given one. I need to test this, but it should work...