Either break or improve pairing
This commit changes the pairing process so that less resources are consumed. It removes the need for a fiber and a listener (~300) bytes.

Instead, a paired driver is created, driver mode is moved to pairing mode, and an additional if statement in onConnect invokes pair() if the driver is in pairing mode.