Fix race with old isconnected code.
The control layer polls the physical layer every 500 ms to see if the bus is / isn't connected. The implementation would reconfigure the pins as a pull up and start GPIO interrupts. If it did this at the wrong time it would interrupt reception of a packet and lock.

FFS!

Now we just check the flags.