Unify flag unsetting in ErrorState & definitely disable irqs if a collision is detected during transmission
Previously flags were being unset before entering the error handler, this commit brings the unsetting to the errorState member function so that upon an error the state is definitely set to a known value.

In addition, when checking for a collision, getDigitalValue was used. There was a bad assumption that a call to this member function would disable gpio irqs which is not the case if the pin is already in digital in mode. There is now a call to setMode(JDSerialMode::off) instead so that gpio irqs are definitely disabled.