Set DEVICE_DAL_VERSION to CODAL_VERSION & update CODAL_VERSION_LONG.
The CODAL_VERSION_LONG macro has been changed to include the
git hash instead of branch name and made sure it is semver
compatible (as long as the target name doesn't contain semver
invalid characters).

The "v" from the version string has also been removed, as it is not
valid semver, and DAL did not add it.