LevelDetectorSPL: Cast `LEVEL_DETECTOR_SPL_8BIT_X_POINT` macros to floats.
This is because setting this value in the codal.json or target.json
files will result in a double being used even if specified as floats.
This is because when the JSON is interpreted via Python and/or CMake
something like `"0.0f"` will still end up in the source code as `0.0`.